refactor: flatten the Python example, dropping the repeated package dir
strategies/sma_crossover/python/sma_crossover/ repeated the strategy name for
no benefit. The inner package existed only to support `python -m sma_crossover`,
which isn't worth a directory level here: each language directory holds exactly
one program, so there is nothing for a package namespace to disambiguate.

    strategies/sma_crossover/python/
      main.py        # entry point, run directly
      strategy.py
      tests/

Run with `python main.py` instead of `python -m sma_crossover`; imports become
plain same-directory imports, which is also what someone dropping these files
into their own project would want. argparse `prog` follows.

Verified on 3.10 and 3.14: pytest, ruff check/format, mypy --strict, CLI help
and validation paths, and the copy-out claim from outside the repository.
