+++ strategies/sma_crossover/python/pyproject.toml
@@ -16,4 +16,4 @@ python_version = "3.10"
 strict = true
 follow_imports = "skip"
 ignore_missing_imports = true
-exclude = ["tests/"]
+exclude = ["tests/", "\\.venv/"]
