--- strategies/sma_crossover/python/sma_crossover/__init__.py
@@ -1,5 +1,0 @@
-"""SMA crossover strategy built on the published Python SDK."""
-
-from .strategy import Signal, evaluate
-
-__all__ = ["Signal", "evaluate"]
