--- strategies/sma_crossover/python/sma_crossover/__main__.py
@@ -1,3 +1,0 @@
-from .main import main
-
-main()
