+++ strategies/sma_crossover/python/strategy.py
